refactor(platform): split auth helpers from main
